# openapi.yaml
openapi: 3.0.1
info:
  title: Rest blog api
  description: Sagger documentaion for my rest blog.
  contact:
    name: Aurel Marishta
    email: aurel.marishta@atis.al
  license:
    name: License of API
  version: API TOS
servers:
- url: http://localhost:8080/
...
paths:
...
  /api/articles:
    get:
      tags:
      - Articles
      summary: View all available articles
      operationId: getAllArticlesUsingGET
      parameters:
      - name: page
        in: query
        description: page
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: size
        schema:
          type: integer
          format: int32
          default: 30
      responses:
        200:
          description: Successfully retrieved list
          content: {}
        401:
          description: You are not authorized to view the resource
          content: {}
        403:
          description: Accessing the resource you were trying to reach is forbidden
          content: {}
        404:
          description: The resource you were trying to reach is not found
          content: {}
      deprecated: false
    post:
      tags:
      - Articles
      summary: Create a new article
      operationId: addArticleUsingPOST
      parameters:
      - name: accountNonExpired
        in: query
        schema:
          type: boolean
      - name: accountNonLocked
        in: query
        schema:
          type: boolean
      - name: authorities[0].authority
        in: query
        schema:
          type: string
      - name: credentialsNonExpired
        in: query
        schema:
          type: boolean
      - name: email
        in: query
        schema:
          type: string
      - name: enabled
        in: query
        schema:
          type: boolean
      - name: firstName
        in: query
        schema:
          type: string
      - name: id
        in: query
        schema:
          type: integer
          format: int64
      - name: lastName
        in: query
        schema:
          type: string
      - name: password
        in: query
        schema:
          type: string
      - name: username
        in: query
        schema:
          type: string
      requestBody:
        description: article
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Successfully created
          content:
            '*/*':
              schema:
                type: object
        401:
          description: Unauthorized
          content: {}
        403:
          description: Authorisation required
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
  /api/articles/{id}:
    get:
      tags:
      - Articles
      summary: View article
      operationId: getArticleUsingGET
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: Successfully retrieved article
          content:
            '*/*':
              schema:
                type: object
        401:
          description: You are not authorized to view the resource
          content: {}
        403:
          description: Accessing the resource you were trying to reach is forbidden
          content: {}
        404:
          description: The resource you were trying to reach is not found
          content: {}
      deprecated: false
    put:
      tags:
      - Articles
      summary: Update article
      operationId: updateArticleUsingPUT
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: newArticle
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleDto'
        required: true
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Created
          content: {}
        202:
          description: Successfully updated
          content:
            '*/*':
              schema:
                type: object
        401:
          description: You are not authorized to update the resource
          content: {}
        403:
          description: Authorisation required
          content: {}
        404:
          description: The resource you were trying to reach is not found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
    delete:
      tags:
      - Articles
      summary: Delete article
      operationId: deleteArticleUsingDELETE
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        204:
          description: Successfully deleted
          content:
            '*/*':
              schema:
                type: object
        401:
          description: You are not authorized to delete the resource
          content: {}
        403:
          description: Authorisation required
          content: {}
        404:
          description: The resource you were trying to reach is not found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
...
  /api/categories:
    get:
      tags:
      - Categories
      summary: getAllArticles
      operationId: getAllArticlesUsingGET_1
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
    post:
      tags:
      - Categories
      summary: addArticle
      operationId: addArticleUsingPOST_1
      requestBody:
        description: categoryDto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Created
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
  /api/categories/{id}:
    put:
      tags:
      - Categories
      summary: updateCategory
      operationId: updateCategoryUsingPUT
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: updatedCategory
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Created
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
  /api/categories/{id}/articles:
    get:
      tags:
      - Categories
      summary: getCategoryArticles
      operationId: getCategoryArticlesUsingGET
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: page
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: size
        schema:
          type: integer
          format: int32
          default: 30
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
...
  /api/tags:
    get:
      tags:
      - Tags
      summary: getAllTags
      operationId: getAllTagsUsingGET
      parameters:
      - name: page
        in: query
        description: page
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: size
        schema:
          type: integer
          format: int32
          default: 30
      responses:
        200:
          description: OK
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
    post:
      tags:
      - Tags
      summary: addPost
      operationId: addPostUsingPOST
      requestBody:
        description: tag
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tag'
        required: true
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        201:
          description: Created
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
  /api/tags/{id}:
    get:
      tags:
      - Tags
      summary: getArticle
      operationId: getArticleUsingGET_1
      parameters:
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
    delete:
      tags:
      - Tags
      summary: deleteTag
      operationId: deleteTagUsingDELETE
      parameters:
      - name: accountNonExpired
        in: query
        schema:
          type: boolean
      - name: accountNonLocked
        in: query
        schema:
          type: boolean
      - name: authorities[0].authority
        in: query
        schema:
          type: string
      - name: credentialsNonExpired
        in: query
        schema:
          type: boolean
      - name: email
        in: query
        schema:
          type: string
      - name: enabled
        in: query
        schema:
          type: boolean
      - name: firstName
        in: query
        schema:
          type: string
      - name: id
        in: path
        description: id
        required: true
        schema:
          type: integer
          format: int64
      - name: id
        in: query
        schema:
          type: integer
          format: int64
      - name: lastName
        in: query
        schema:
          type: string
      - name: password
        in: query
        schema:
          type: string
      - name: username
        in: query
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        204:
          description: No Content
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
      deprecated: false
      security:
      - bearerAuth: []
  /api/tags/{tag}/articles:
    get:
      tags:
      - Tags
      summary: getAllArticlesWithTag
      operationId: getAllArticlesWithTagUsingGET
      parameters:
      - name: tag
        in: path
        description: tag
        required: true
        schema:
          type: string
      responses:
        200:
          description: OK
          content:
            '*/*':
              schema:
                type: object
        401:
          description: Unauthorized
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
      deprecated: false
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    Article:
      title: Article
      type: object
      properties:
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        featuredImage:
          $ref: '#/components/schemas/Image'
        id:
          type: integer
          format: int64
        metas:
          type: array
          items:
            $ref: '#/components/schemas/ArticleMeta'
        open_new_window:
          type: boolean
        position:
          type: integer
          format: int64
        reactions:
          type: array
          items:
            $ref: '#/components/schemas/Reaction'
        scheduledAt:
          type: string
          format: date-time
        slug:
          type: string
        summary:
          type: string
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
...
    Image:
      title: Image
      type: object
      properties:
        alt:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        description:
          type: string
        id:
          type: integer
          format: int64
        name:
          type: string
        size:
          type: integer
          format: int64
        type:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
        url:
          type: string
    ArticleDto:
      title: ArticleDto
      type: object
      properties:
        _featured:
          type: boolean
        authorised:
          type: boolean
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        enabled:
          type: boolean
        featuredImage:
          $ref: '#/components/schemas/Image'
        images:
          type: array
          items:
            type: integer
            format: int64
        open_new_window:
          type: boolean
        position:
          type: integer
          format: int64
        scheduledAt:
          type: string
        slug:
          type: string
        summary:
          type: string
        tags:
          type: array
          items:
            type: integer
            format: int64
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        user:
          type: integer
          format: int64
    ArticleMeta:
      title: ArticleMeta
      type: object
      properties:
        id:
          type: integer
          format: int32
        metaKey:
          type: string
        metaValue:
          type: string
...
    Reaction:
      title: Reaction
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
        reaction:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: integer
          format: int64
        user:
          type: object
          properties: {}
    CategoryDto:
      title: CategoryDto
      type: object
      properties:
        articles:
          type: array
          items:
            type: integer
            format: int64
        name:
          type: string
        parent:
          type: integer
          format: int64
        subCategories:
          type: array
          items:
            type: integer
            format: int64
...
    Tag:
      title: Tag
      type: object
      properties:
        createdAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        tag:
          type: string
        updatedAt:
          type: string
          format: date-time
...